Keep checked-cast ratchet issue-scoped
